backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "mysql-connector-python>=9.3.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.10.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
]
